more generalization in packer, packer now generates ubuntu from imgbuilder.preseed

+++ roles/imgbuilder/packer/templates/linux_script/vagrant.sh
@@ -15,10 +15,17 @@ fi
 echo "Set up sudo for vagrant"
 echo "${VAGRANT_USER}        ALL=(ALL)       NOPASSWD: ALL" >> /etc/sudoers
 
-echo "Install vagrant keys"
+echo "Install vagrant keys (if not already a .ssh/authorized_keys file present)"
 mkdir $VAGRANT_HOME/.ssh
 chmod 700 $VAGRANT_HOME/.ssh
 cd $VAGRANT_HOME/.ssh
-wget --no-check-certificate "${VAGRANT_KEY_URL}" -O authorized_keys
+
+if test -f authorized_keys; then
+    echo "aborted, there is already a authorized_keys file with content:"
+    cat authorized_keys
+else
+    wget --no-check-certificate "${VAGRANT_KEY_URL}" -O authorized_keys
+fi
+
 chmod 600 $VAGRANT_HOME/.ssh/authorized_keys
 chown -R $VAGRANT_USER:$VAGRANT_USER $VAGRANT_HOME/.ssh
